Move TupleMatchers class to org.elasticsearch.test.hamcrest package (#51359)
